remove util file
